+++ CHANGELOG.md
@@ -12,3 +12,4 @@
 - Warn when using Java 9 or newer that some packs may not work
 - Allow packs to set min/max java versions for their packs
 - Allow packs to disable edit mods button
+- Add browse button to java path box